% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stuff.R
\name{add_stuff}
\alias{add_stuff}
\title{Add stuff
 @param a A number
 @param b Another number
 @return Yet another number
 @examples
     add_stuff(3,2)
 @export}
\usage{
add_stuff(a, b)
}
\description{
Add stuff
 @param a A number
 @param b Another number
 @return Yet another number
 @examples
     add_stuff(3,2)
 @export
}
