% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stuff.R
\name{add_stuff}
\alias{add_stuff}
\title{Add stuff}
\usage{
add_stuff(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{Another number}
}
\value{
Yet another number
}
\description{
Add stuff
}
\examples{
     add_stuff(3,2)
}
